Move segment code out of tools
So it's available without Gtk.
