+++ gaphor/diagram/__init__.py
@@ -10,3 +10,4 @@ gi.require_version("Pango", "1.0")
 gi.require_version("PangoCairo", "1.0")
 
 import gaphor.diagram._connector
+import gaphor.diagram.segment
