+++ gaphor/diagram/segment.py
