+++ gaphor/diagram/tests/test_segment.py
@@ -1,7 +1,7 @@
 from gaphas.segment import Segment
 
 from gaphor.diagram.presentation import LinePresentation
-from gaphor.diagram.tools.segment import PresentationSegment
+from gaphor.diagram.segment import PresentationSegment
 
 
 def test_presentation_segment_is_selected(diagram):
